fix(ui): remediate cross-platform UI/UX configuration and accessibility
- Consolidate missing UI primitives (DropdownMenu, Accordion, Card,
  Badge, Skeleton) into @esparex/ui.
- Update apps/web/src/components/ui local primitives to re-exports.
- Replace hardcoded slate/white utilities in Header, MobileBottomNav,
  and BottomActionsBar with semantic design tokens.
- Bind mobile ThemeProvider to react-native useColorScheme.
- Expand small button touch target tap area (>=44px) in Button and
  AppButton.
- Replace static bottom padding with dynamic useSafeAreaInsets in
  MarketplaceScreen.
- Connect Input and Textarea to FieldContext for auto WCAG
  aria-invalid and aria-describedby linking.
- Wrap native PostAdScreen in KeyboardAvoidingView to prevent virtual
  keyboard obscuration.
- Replace h-[100vh] with h-[100dvh] in AccountMessagesWorkspace and
  not-found.tsx.
- Add inert attribute to closed MobileNavDrawer to isolate focus.

Closes #118
