+++ apps/mobile/src/features/listings/presentation/screens/MarketplaceScreen.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useCallback, useMemo } from 'react';
 import { FlatList, RefreshControl, View, StyleSheet } from 'react-native';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Screen, Container } from '@esparex/mobile-ui';
 import { ListingQueryParams } from '@esparex/contracts';
 import { useListings } from '../hooks/useListings';
@@ -14,6 +15,7 @@ import { ROUTES } from '../../../../navigation/routes';
 import { Listing } from '../../domain/Listing';
 
 export const MarketplaceScreen = () => {
+  const insets = useSafeAreaInsets();
   const [filters, setFilters] = useState<ListingQueryParams>({});
   const [isFilterModalOpen, setIsFilterModalOpen] = useState(false);
 
@@ -107,7 +109,7 @@ export const MarketplaceScreen = () => {
             data={listings}
             renderItem={renderItem}
             keyExtractor={keyExtractor}
-            contentContainerStyle={styles.scrollContent}
+            contentContainerStyle={{ padding: 16, paddingBottom: insets.bottom + 64 }}
             showsVerticalScrollIndicator={false}
             removeClippedSubviews={true}
             windowSize={5}
@@ -159,6 +161,3 @@ export const MarketplaceScreen = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  scrollContent: { padding: 16, paddingBottom: 100 },
-});
