+++ apps/web/src/components/ui/card.tsx
@@ -1,67 +1,13 @@
-import * as React from "react";
+"use client";
 
-import { cn } from "./utils";
-
-function Card({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      data-slot="card"
-      className={cn(
-        "bg-white text-slate-800 flex flex-col gap-3 rounded-xl border border-slate-200/70 shadow-xs hover:border-slate-300/80 transition-colors",
-        className,
-      )}
-      {...props}
-    />
-  );
-}
-
-function CardHeader({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      data-slot="card-header"
-      className={cn(
-        "@container/card-header grid auto-rows-min grid-rows-[auto_auto] items-start gap-1 px-3.5 pt-3.5 sm:px-4 sm:pt-4 has-data-[slot=card-action]:grid-cols-[1fr_auto] [.border-b]:pb-2 sm:[.border-b]:pb-3",
-        className,
-      )}
-      {...props}
-    />
-  );
-}
-
-function CardTitle({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <h4
-      data-slot="card-title"
-      className={cn("leading-snug font-semibold text-h4 text-slate-800 tracking-tight", className)}
-      {...props}
-    />
-  );
-}
-
-function CardDescription({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <p
-      data-slot="card-description"
-      className={cn("text-slate-500 text-caption font-normal", className)}
-      {...props}
-    />
-  );
-}
-
-function CardContent({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      data-slot="card-content"
-      className={cn("px-3.5 sm:px-4 [&:last-child]:pb-3.5 sm:[&:last-child]:pb-4", className)}
-      {...props}
-    />
-  );
-}
-
+/**
+ * card.tsx — Single Source of Truth Re-export
+ * Re-exports canonical Card primitives from @esparex/ui.
+ */
 export {
   Card,
   CardHeader,
   CardTitle,
   CardDescription,
   CardContent,
-};
+} from "@esparex/ui";
