+++ packages/ui/src/atoms/Button.tsx
@@ -5,7 +5,7 @@ import { Slot } from "@radix-ui/react-slot";
 import { cn } from "../utils";
 
 const base =
-  "inline-flex items-center justify-center gap-2 rounded-xl text-sm transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0 disabled:pointer-events-none disabled:opacity-50 relative before:absolute before:inset-y-[-2px] before:inset-x-0 before:content-[''] before:pointer-events-auto";
+  "inline-flex items-center justify-center gap-2 rounded-xl text-sm transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0 disabled:pointer-events-none disabled:opacity-50 relative before:absolute before:inset-y-[-4px] before:inset-x-0 before:content-[''] before:pointer-events-auto";
 
 const variants = {
   default: "bg-primary text-primary-foreground font-semibold hover:bg-primary/90",
