+++ packages/ui/src/feedback/Skeleton.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+import { cn } from "../utils";
+
+function Skeleton({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) {
+  return (
+    <div
+      className={cn("animate-pulse rounded-md bg-muted", className)}
+      {...props}
+    />
+  );
+}
+
+export { Skeleton };
